requirements.txt
streamlit>=1.28.0
pdfplumber>=0.10.0
Pillow>=10.0.0
groq>=0.9.0
mistralai>=1.0.0
pandas>=2.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.23.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
plotly>=5.18.0
